Add separate test class for DoctrineKeyValueStyleRule strict mode checks
