+++ tests/rules/DoctrineKeyValueStyleRuleTest.php
@@ -6,8 +6,6 @@ namespace staabm\PHPStanDba\Tests;
 
 use PHPStan\Rules\Rule;
 use PHPStan\Testing\RuleTestCase;
-use staabm\PHPStanDba\QueryReflection\QueryReflection;
-use staabm\PHPStanDba\QueryReflection\RuntimeConfiguration;
 use staabm\PHPStanDba\Rules\DoctrineKeyValueStyleRule;
 
 /**
@@ -15,11 +13,6 @@ use staabm\PHPStanDba\Rules\DoctrineKeyValueStyleRule;
  */
 class DoctrineKeyValueStyleRuleTest extends RuleTestCase
 {
-    protected function tearDown(): void
-    {
-        QueryReflection::getRuntimeConfiguration()->parameterTypeValidation(RuntimeConfiguration::VALIDATION_MODE_LAX);
-    }
-
     protected function getRule(): Rule
     {
         return self::getContainer()->getByType(DoctrineKeyValueStyleRule::class);
@@ -32,7 +25,7 @@ class DoctrineKeyValueStyleRuleTest extends RuleTestCase
         ];
     }
 
-    public function testSyntaxErrorInMethod(): void
+    public function testRule(): void
     {
         $expectedErrors = [
             [
@@ -78,23 +71,5 @@ class DoctrineKeyValueStyleRuleTest extends RuleTestCase
         ];
 
         $this->analyse([__DIR__ . '/data/doctrine-key-value-style.php'], $expectedErrors);
-    }
-
-    public function testIntegerRanges(): void
-    {
-        $expectedErrors = [
-            [
-                'Query error: Column "ada.adaid" expects value type int<-32768, 32767>, got type int',
-                10,
-            ],
-            [
-                'Query error: Column "ada.adaid" expects value type int<-32768, 32767>, got type int<0, 65535>',
-                18,
-            ],
-        ];
-
-        QueryReflection::getRuntimeConfiguration()->parameterTypeValidation(RuntimeConfiguration::VALIDATION_MODE_STRICT);
-
-        $this->analyse([__DIR__ . '/data/doctrine-key-value-style-integer-ranges.php'], $expectedErrors);
     }
 }
